requirements.txt
asgiref==3.2.3
astroid==2.3.3
cffi==1.14.0
colorama==0.4.3
cryptography==2.8
Django==3.0.3
django-appconf==1.0.3
django-cors-headers==3.2.1
django-cryptography==1.0
django-rest-framework-nested==0.0.1
django-tinymce==2.8.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-nested-routers==0.91
entrypoints==0.3
flake8==3.7.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pep8==1.7.1
Pillow==7.0.0
pycodestyle==2.5.0
pycparser==2.20
pycryptodome==3.9.7
pyflakes==2.1.1
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
wrapt==1.11.2
